% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bmygibbsC}
\alias{bmygibbsC}
\title{A gibbs sampler using Rcpp}
\usage{
bmygibbsC(N, rho, mu1, mu2, sigma1, sigma2)
}
\arguments{
\item{N}{Number of random numbers}

\item{rho}{Correlation coefficient of two-dimensional normal distribution}

\item{mu1}{Mean value of the first normal distribution}

\item{mu2}{Mean value of the second normal distribution}

\item{sigma1}{Standard deviation of the first normal distribution}

\item{sigma2}{Standard deviation of the second normal distribution}
}
\value{
Returns chain
}
\description{
Implement A gibbs sampler using Rcpp for generating the 
two-dimensional normal distribution
}
\examples{
\dontrun{
bmygibbsC(1000,0.7,0,0,1,1)
}
}
