% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmykmeans.R
\name{bmykmeans}
\alias{bmykmeans}
\title{K-mean cluster using R}
\usage{
bmykmeans(dataset, k)
}
\arguments{
\item{dataset}{a data matrix}

\item{k}{the number of cluster}
}
\value{
Cluster center for each class and the cluster name of each point
}
\description{
Calculate the center of K-mean cluster using R and return the cluster of each point
}
\examples{
\dontrun{
dataset<-iris[,1:2]
k<-3
z<-bmykmeans(dataset,k)
z
}
}
