% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmymixnorsam.R
\name{bmymixnorsam}
\alias{bmymixnorsam}
\title{Generate random numbers related to normal distribution using R}
\usage{
bmymixnorsam(mode, burn, p, mu1, sigma1, mu2, sigma2, mu0, sigma0, m, rho)
}
\arguments{
\item{mode}{1 is to choose to mix two normal distributions, 2 is to choose to generate two-dimensional normal distributions}

\item{burn}{Points removed before convergence}

\item{p}{Proportion of the first normal distribution in mode1}

\item{mu1}{Mean value of the first normal distribution}

\item{sigma1}{Standard deviation of the first normal distribution}

\item{mu2}{Mean value of the second normal distribution}

\item{sigma2}{Standard deviation of the second normal distribution}

\item{mu0}{Mean of the proposed distribution in model 1}

\item{sigma0}{Standard deviation of the proposed distribution in model 1}

\item{m}{the the length of the chain}

\item{rho}{Correlation coefficient of two-dimensional normal distribution in model 2}
}
\value{
the relevant convergence and scatter plot
}
\description{
This function can generate a mixed distribution of two normal distributions in mode 1 and a two-dimensional normal distribution in mode 2 Mode 1 adopts independent sampling algorithm with normal distribution as the  proposed distribution, and mode 2 adopts gibbs sampling algorithm.
}
\examples{
\dontrun{
bmymixnorsam(1,1000,0.2,0,1,5,1,4,1,10000,0.2)
bmymixnorsam(2,1000,0.2,0,1,2,1,0,0,10000,-0.7)
}
}
