\name{StatComp22013-package}
\alias{StatComp22013-package}
\docType{package}
\title{
  This package provides the final homework of statistical computing.
}
\description{
  The package first contains the job summary of statistical calculation, and then contains two R language functions. The first one is the implementation of kmeans clustering method, and the second one is the generation of random numbers subject to mixed normal or two-dimensional normal. The package also contains a C language function, which is about gibbs sampling.
}
\author{
Chunxi Liu

Maintainer: Chunxi Liu <liuchunxi@mail.ustc.edu.cn>
}
\keyword{ package }